+++ resources/reacts/tw.json
@@ -29,12 +29,12 @@
     {
       "match_whole_word": false,
       "keywords": [
-        "^(?:台|臺)灣",
-        "^(?:台|臺)北",
+        "(?:台|臺)灣",
+        "(?:台|臺)北",
         "新北",
         "桃園",
-        "^(?:台|臺)中",
-        "^(?:台|臺)南",
+        "(?:台|臺)中",
+        "(?:台|臺)南",
         "高雄",
         "新竹",
         "苗栗",
@@ -45,7 +45,7 @@
         "屏東",
         "宜蘭",
         "花蓮",
-        "^(台|臺)東",
+        "(台|臺)東",
         "澎湖",
         "金門",
         "連江",
